+++ app/services/task/filter/wildcard_search.rb
@@ -3,12 +3,30 @@ class Task::Filter::WildcardSearch < Task::Filter::Base
     if filters[:wildcard_search] != 'null' && filters[:wildcard_search].present?
       @tasks = tasks
       @filters = filters
-      tasks = tasks.where('activities.subject ilike ? OR activities.id IN (?)', "%#{filters[:wildcard_search]}%", tagged_tasks_ids)
+      tasks = tasks.where('activities.subject ilike ? OR activities.id IN (?)', wildcard_string, relevant_task_ids)
     end
     tasks
   end
 
-  def tagged_tasks_ids
-    @tasks.tagged_with(@filters[:wildcard_search].split(',').flatten, any: @filters[:wildcard_search] == 'true').ids
+  private
+
+  def relevant_task_ids
+    (tagged_task_ids + task_ids_with_relevant_contact_name + task_ids_with_relevant_comment).uniq
+  end
+
+  def tagged_task_ids
+    @tasks.tagged_with(@filters[:wildcard_search].split(',').flatten, wild: true, any: true).ids
+  end
+
+  def task_ids_with_relevant_contact_name
+    @tasks.joins(:contacts).where('contacts.name ilike ?', wildcard_string).ids
+  end
+
+  def task_ids_with_relevant_comment
+    @tasks.joins(:comments).where('activity_comments.body ilike ?', wildcard_string).ids
+  end
+
+  def wildcard_string
+    "%#{@filters[:wildcard_search]}%"
   end
 end