+++ app/workers/mail_chimp/primary_list_sync_worker.rb
@@ -3,6 +3,6 @@ class MailChimp::PrimaryListSyncWorker
   sidekiq_options queue: :api_mail_chimp_sync_worker, unique: :until_executed
 
   def perform(mail_chimp_account)
-    MailChimp::Syncer.new(mail_chimp_account).sync_with_primary_list
+    MailChimp::Syncer.new(mail_chimp_account).two_way_sync_with_primary_list
   end
 end